move helper function to class method
