+++ commcare_connect/organization/tests/test_models.py
@@ -1,16 +1,33 @@
 from datetime import timedelta
 
 import pytest
+from django.contrib.auth.models import Permission
 from django.db import IntegrityError
 from django.utils import timezone
 
-from commcare_connect.organization.models import LLOEntity, OrganizationInvite, UserOrganizationMembership
+from commcare_connect.organization.models import (
+    LLOEntity,
+    Organization,
+    OrganizationInvite,
+    UserOrganizationMembership,
+)
+from commcare_connect.users.models import User
 from commcare_connect.users.tests.factories import (
     MembershipFactory,
     OrganizationFactory,
     OrganizationInviteFactory,
     UserFactory,
 )
+
+
+def _user_with_privilege(privilege: str | None, own_org: Organization) -> User:
+    user = UserFactory(is_superuser=privilege == "superuser")
+    if privilege == "member":
+        MembershipFactory(user=user, organization=own_org)
+    if privilege == "permission":
+        user.user_permissions.add(Permission.objects.get(codename="workspace_entity_management_access"))
+        user = User.objects.get(pk=user.pk)  # drop the cached permissions
+    return user
 
 
 class TestLLOEntity:
@@ -62,6 +79,21 @@ class TestOrganization:
         if expect_viewer:
             expected.append(viewer.email)
         assert sorted(emails) == sorted(expected)
+
+    @pytest.mark.parametrize(
+        "privilege, visible",
+        [
+            (None, set()),
+            ("member", {"own"}),
+            ("permission", {"own", "other"}),
+            ("superuser", {"own", "other"}),
+        ],
+    )
+    def test_visible_to(self, privilege, visible):
+        orgs = {"own": OrganizationFactory(), "other": OrganizationFactory()}
+        user = _user_with_privilege(privilege, orgs["own"])
+
+        assert set(Organization.visible_to(user)) == {orgs[key] for key in visible}
 
 
 @pytest.mark.django_db
